Refactor data out
