+++ flat-state/src/lib.rs
@@ -1,2 +1,3 @@
+mod data;
 pub mod filter;
 pub mod state;
