+++ primitives/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod block_with_tx_hash;
 pub mod types;
+pub mod utils;
 
 pub use near_indexer_primitives;
 pub use near_primitives;
